requirements.txt
Flask==2.3.3
Flask-Bootstrap4==4.0.2
supabase==1.0.3
python-dateutil==2.8.2
Werkzeug==2.3.7
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
requests==2.31.0
urllib3==2.0.4
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
